bugfix(PDF): Removed a NaN generation
Very rarely, a random direction can generate NaN due to float imprecision, since we can generate directions that are very close to orthogonal to our normal. This is fixed by ensuring we never return a null pdf. This might not be ideal but i believe it is the best course of action for now.